ASTUTE

Perform the ASTUTE framework to estimate K expression signatures associated to a set of somatic alterations provided as input.

ASTUTE(
  alterations,
  expression,
  regularization = TRUE,
  nboot = 100,
  num_processes = Inf,
  verbose = TRUE
)

Arguments

alterations

Input binary alterations matrix.

expression

Input log2(x+1)-transformed normalized expression matrix.

regularization

Boolean. If TRUE, perform regularization via elastic net with LASSO penalty.

nboot

Number of bootstrap sampling (minimum 2) to be performed for a robust estimation of the expression signatures. If bootstrap does not need to be executed, this parameter needs to be set to either NA or NULL.

num_processes

Number of processes to be used during parallel execution. To execute in single process mode, this parameter needs to be set to either NA or NULL.

verbose

Boolean. If TRUE, print status messages.

Value

A list with the discovered expression signatures. It includes 7 elements: input_data: list providing the input data (i.e., alterations and expression data); bootstrap: results of the inference by bootstrap (i.e., alpha alterations matrix, beta matrix, and intercept estimates); parameters: list with the paremeters used for the inference (i.e., regularization TRUE/FALSE and nboot); goodness_fit: goodness of fit estimated as the cosine similarity comparing observations and predictions; fold_changes: log2 fold changes estimates; pvalues: p-values estimates; qvalues: p-values estimates corrected for false discovery.

Examples

set.seed(12345)
data(datasetExample)
resExample <- ASTUTE( alterations = datasetExample$alterations, 
                      expression = datasetExample$expression, 
                      regularization = TRUE, 
                      nboot = NA, 
                      num_processes = NA, 
                      verbose = FALSE )